Fix warning in GOrgueWave (MK)
git-svn-id: https://svn.code.sf.net/p/ourorgan/svn/trunk@273 69509848-2a73-49d5-a893-c4d8c934025a

+++ src/grandorgue/GOrgueWave.cpp
@@ -258,7 +258,7 @@ void GOrgueWave::Open(const wxString& filename)
 		/* This is a bit more leaniant than the original code... it will
 		 * truncate the usable size of the file if the size on disk is larger
 		 * than the size of the RIFF chunk */
-		if (length > riffChunkSize + 8)
+		if ((unsigned long)length > riffChunkSize + 8)
 			length = riffChunkSize + 8;
 
 		/* Find required chunks... */
